poc/requirements.txt
# LangGraph POC Dependencies (Local Version)
langgraph>=0.2.0
langchain>=0.1.0
langchain-openai>=0.0.5
pyyaml>=6.0
python-dotenv>=1.0.0
pydantic>=2.5.0
langchain-aws>=0.1.0
boto3>=1.26.0